refactor(cognition): decompose act_observe.rs into a directory module (recency channel extracted)
Pure code-motion, behavior-identical. act_observe.rs (3,211 lines) becomes a
directory module: act_observe/mod.rs (3,011) + act_observe/recency.rs (215).

The six recency-channel helpers — turning one executed act (args + result) into
bounded, legible working-memory text — were one responsibility buried in a 3k-line
file. They now live in their own single-responsibility submodule. truncate_chars
and summarize_args_for_recall are recency-internal (private); the four with parent/
test callers are pub(super) and re-exported so every call site stays unqualified.
No public API change, no logic change.

Extracted via a deterministic splitter with boundary-drift guards (fails loud if
the line anchors move), not fragile hand-editing. cargo check green; the 28
act_observe module tests pass unchanged.

First cut of the continuous-decomposition discipline: 20 more files sit over 1,800
lines (eval 4,740; serving_daemon 4,618; service_loop 4,438). Every file I touch
leaves smaller.
